+++ packages/verified-fetch/src/verified-fetch.ts
@@ -26,7 +26,6 @@ import { selectOutputType } from './utils/select-output-type.js'
 import { serverTiming } from './utils/server-timing.js'
 import type { CIDDetail, ContentTypeParser, CreateVerifiedFetchOptions, Resource, ResourceDetail, VerifiedFetchInit as VerifiedFetchOptions } from './index.js'
 import type { VerifiedFetchPlugin, PluginContext, PluginOptions } from './plugins/types.js'
-import type { RequestFormatShorthand } from './types.js'
 import type { Helia, SessionBlockstore } from '@helia/interface'
 import type { Blockstore } from 'interface-blockstore'
 
@@ -54,16 +53,6 @@ function convertOptions (options?: VerifiedFetchOptions): (Omit<VerifiedFetchOpt
     ...options,
     signal
   }
-}
-
-// TODO: merge/combine with PluginContext
-interface FinalResponseContext {
-  cid?: ParsedUrlStringResults['cid']
-  reqFormat?: RequestFormatShorthand
-  ttl?: ParsedUrlStringResults['ttl']
-  protocol?: ParsedUrlStringResults['protocol']
-  ipfsPath?: string
-  query?: ParsedUrlStringResults['query']
 }
 
 export class VerifiedFetch {
@@ -162,7 +151,7 @@ export class VerifiedFetch {
    * Server-Timing header to the response if it has been collected. It should be used for any final processing of the
    * response before it is returned to the user.
    */
-  private handleFinalResponse (response: Response, { query, cid, reqFormat, ttl, protocol, ipfsPath }: FinalResponseContext = {}): Response {
+  private handleFinalResponse (response: Response, { query, cid, reqFormat, ttl, protocol, ipfsPath, pathDetails }: Partial<PluginContext> = {}): Response {
     if (this.serverTimingHeaders.length > 0) {
       const headerString = this.serverTimingHeaders.join(', ')
       response.headers.set('Server-Timing', headerString)
@@ -199,7 +188,7 @@ export class VerifiedFetch {
     }
 
     if (cid != null && response.headers.get('etag') == null) {
-      response.headers.set('etag', getETag({ cid, reqFormat, weak: false }))
+      response.headers.set('etag', getETag({ cid: pathDetails?.terminalElement.cid ?? cid, reqFormat, weak: false }))
     }
 
     if (protocol != null) {
@@ -308,22 +297,11 @@ export class VerifiedFetch {
     const withServerTiming = options?.withServerTiming ?? this.withServerTiming
 
     options?.onProgress?.(new CustomProgressEvent<ResourceDetail>('verified-fetch:request:start', { resource }))
-    // resolve the CID/path from the requested resource
-    let cid: ParsedUrlStringResults['cid']
-    let path: ParsedUrlStringResults['path']
-    let query: ParsedUrlStringResults['query']
-    let ttl: ParsedUrlStringResults['ttl']
-    let protocol: ParsedUrlStringResults['protocol']
-    let ipfsPath: string
+
+    let parsedResult: ParsedUrlStringResults
     try {
-      const result = await this.handleServerTiming('parse-resource', '', async () => parseResource(resource, { ipns: this.ipns, logger: this.helia.logger }, { withServerTiming, ...options }), withServerTiming)
-      cid = result.cid
-      path = result.path
-      query = result.query
-      ttl = result.ttl
-      protocol = result.protocol
-      ipfsPath = result.ipfsPath
-      this.serverTimingHeaders.push(...result.serverTimings.map(({ header }) => header))
+      parsedResult = await this.handleServerTiming('parse-resource', '', async () => parseResource(resource, { ipns: this.ipns, logger: this.helia.logger }, { withServerTiming, ...options }), withServerTiming)
+      this.serverTimingHeaders.push(...parsedResult.serverTimings.map(({ header }) => header))
     } catch (err: any) {
       options?.signal?.throwIfAborted()
       this.log.error('error parsing resource %s', resource, err)
@@ -331,11 +309,11 @@ export class VerifiedFetch {
       return this.handleFinalResponse(badRequestResponse(resource.toString(), err))
     }
 
-    options?.onProgress?.(new CustomProgressEvent<CIDDetail>('verified-fetch:request:resolve', { cid, path }))
+    options?.onProgress?.(new CustomProgressEvent<CIDDetail>('verified-fetch:request:resolve', { cid: parsedResult.cid, path: parsedResult.path }))
 
-    const acceptHeader = getResolvedAcceptHeader({ query, headers: options?.headers, logger: this.helia.logger })
+    const acceptHeader = getResolvedAcceptHeader({ query: parsedResult.query, headers: options?.headers, logger: this.helia.logger })
 
-    const accept: string | undefined = selectOutputType(cid, acceptHeader)
+    const accept: string | undefined = selectOutputType(parsedResult.cid, acceptHeader)
     this.log('output type %s', accept)
 
     if (acceptHeader != null && accept == null) {
@@ -344,30 +322,28 @@ export class VerifiedFetch {
 
     const responseContentType: string = accept?.split(';')[0] ?? 'application/octet-stream'
 
-    const redirectResponse = await getRedirectResponse({ resource, options, logger: this.helia.logger, cid })
+    const redirectResponse = await getRedirectResponse({ resource, options, logger: this.helia.logger, cid: parsedResult.cid })
     if (redirectResponse != null) {
       return this.handleFinalResponse(redirectResponse)
     }
 
-    const context: PluginContext = { cid, path, resource: resource.toString(), accept, query, options, withServerTiming, onProgress: options?.onProgress, modified: 0 }
+    const context: PluginContext = {
+      ...parsedResult,
+      resource: resource.toString(),
+      accept,
+      options,
+      withServerTiming,
+      onProgress: options?.onProgress,
+      modified: 0
+    }
 
-    this.log.trace('finding handler for cid code "%s" and response content type "%s"', cid.code, responseContentType)
+    this.log.trace('finding handler for cid code "%s" and response content type "%s"', parsedResult.cid.code, responseContentType)
 
     const response = await this.runPluginPipeline(context)
 
-    options?.onProgress?.(new CustomProgressEvent<CIDDetail>('verified-fetch:request:end', { cid, path }))
+    options?.onProgress?.(new CustomProgressEvent<CIDDetail>('verified-fetch:request:end', { cid: parsedResult.cid, path: parsedResult.path }))
 
-    return this.handleFinalResponse(response ?? notSupportedResponse(resource.toString()), {
-      query: {
-        ...query,
-        ...context.query
-      },
-      cid,
-      reqFormat: context.reqFormat,
-      ttl,
-      protocol,
-      ipfsPath
-    })
+    return this.handleFinalResponse(response ?? notSupportedResponse(resource.toString()), context)
   }
 
   /**